byteapp/backend/requirements.txt
flask
flask-cors
gunicorn
pymongo
python-dotenv
google-generativeai
PyPDF2
python-docx
